+++ threedi_api_client/aio/files.py
@@ -19,8 +19,15 @@ CONTENT_RANGE_REGEXP = re.compile(r"^bytes (\d+)-(\d+)/(\d+|\*)$")
 RETRY_STATUSES = frozenset({413, 429, 503})  # like in urllib3
 DEFAULT_CONN_LIMIT = 4  # for downloads only (which are parrallel)
 # only timeout on the socket, not on Python code (like urllib3)
-DEFAULT_TIMEOUT = aiohttp.ClientTimeout(total=None, sock_connect=5.0, sock_read=5.0)
-
+DEFAULT_DOWNLOAD_TIMEOUT = aiohttp.ClientTimeout(
+    total=None, sock_connect=5.0, sock_read=5.0
+)
+# Default upload timeout has an increased socket read timeout, because MinIO
+# takes very long for completing the upload for larger files. The limit of 10 minutes
+# should accomodate files up to 150 GB.
+DEFAULT_UPLOAD_TIMEOUT = aiohttp.ClientTimeout(
+    total=None, sock_connect=5.0, sock_read=600.0
+)
 
 logger = logging.getLogger(__name__)
 
@@ -61,7 +68,7 @@ async def download_file(
     executor: Optional[ThreadPoolExecutor] = None,
     retries: int = 3,
     backoff_factor: float = 1.0,
-    callback_func: Optional[Callable[[int, int], Awaitable[None]]] = None
+    callback_func: Optional[Callable[[int, int], Awaitable[None]]] = None,
 ) -> Tuple[Path, int]:
     """Download a file to a specified path on disk.
 
@@ -116,7 +123,7 @@ async def download_file(
                 connector=connector,
                 retries=retries,
                 backoff_factor=backoff_factor,
-                callback_func=callback_func
+                callback_func=callback_func,
             )
     except Exception:
         # Clean up a partially downloaded file
@@ -165,7 +172,7 @@ async def download_fileobj(
     connector: Optional[aiohttp.BaseConnector] = None,
     retries: int = 3,
     backoff_factor: float = 1.0,
-    callback_func: Optional[Callable[[int, int], Awaitable[None]]] = None
+    callback_func: Optional[Callable[[int, int], Awaitable[None]]] = None,
 ) -> int:
     """Download a url to a file object using multiple requests.
 
@@ -209,7 +216,7 @@ async def download_fileobj(
     # servers support that (e.g. Minio).
     request_kwargs = {
         "url": url,
-        "timeout": DEFAULT_TIMEOUT if timeout is None else timeout,
+        "timeout": DEFAULT_DOWNLOAD_TIMEOUT if timeout is None else timeout,
         "retries": retries,
         "backoff_factor": backoff_factor,
     }
@@ -250,7 +257,11 @@ async def download_fileobj(
                     )
                 )
                 if callable(callback_func):
-                    total: int = file_size if (i + 1) * chunk_size - 1 > file_size else (i + 1) * chunk_size - 1
+                    total: int = (
+                        file_size
+                        if (i + 1) * chunk_size - 1 > file_size
+                        else (i + 1) * chunk_size - 1
+                    )
                     await callback_func(total, file_size)
         except Exception:
             # in case of an exception, cancel all tasks
@@ -271,7 +282,7 @@ async def upload_file(
     executor: Optional[ThreadPoolExecutor] = None,
     retries: int = 3,
     backoff_factor: float = 1.0,
-    callback_func: Optional[Callable[[int, int], Awaitable[None]]] = None
+    callback_func: Optional[Callable[[int, int], Awaitable[None]]] = None,
 ) -> int:
     """Upload a file at specified file path to a url.
 
@@ -284,7 +295,8 @@ async def upload_file(
         chunk_size: The size of the chunk in the streaming upload. Note that this
             function does not do multipart upload. Default: 16MB.
         timeout: The total timeout of the upload in seconds.
-            By default, there is no total timeout, but only socket timeouts of 5s.
+            By default, there is no total timeout, but only socket connect timeout
+            of 5 seconds and a socket read timeout of 10 minutes.
         connector: An optional aiohttp connector to support connection pooling.
         md5: The MD5 digest (binary) of the file. Supply the MD5 if you already
             have access to it. Otherwise this function will compute it for you.
@@ -323,13 +335,17 @@ async def upload_file(
             executor=executor,
             retries=retries,
             backoff_factor=backoff_factor,
-            callback_func=callback_func
+            callback_func=callback_func,
         )
 
     return size
 
 
-async def _iter_chunks(fileobj, chunk_size: int, callback_func: Optional[Callable[[int], Awaitable[None]]] = None):
+async def _iter_chunks(
+    fileobj,
+    chunk_size: int,
+    callback_func: Optional[Callable[[int], Awaitable[None]]] = None,
+):
     """Yield chunks from a file stream"""
     assert chunk_size > 0
     uploaded_bytes: int = 0
@@ -361,14 +377,21 @@ async def _compute_md5(
     return await loop.run_in_executor(executor, hasher.digest)
 
 
-async def _upload_request(client, fileobj, chunk_size, callback_func: Optional[Callable[[int, int], Awaitable[None]]],  *args, **kwargs):
+async def _upload_request(
+    client,
+    fileobj,
+    chunk_size,
+    callback_func: Optional[Callable[[int, int], Awaitable[None]]],
+    *args,
+    **kwargs
+):
     """Send a request with the contents of fileobj as iterable in the body"""
     file_size: int = await fileobj.seek(0, 2)
 
     await fileobj.seek(0)
 
     async def callback(uploaded_bytes: int):
-        if callable(callback_func):    
+        if callable(callback_func):
             if uploaded_bytes > file_size:
                 uploaded_bytes = file_size
             await callback_func(uploaded_bytes, file_size)
@@ -390,7 +413,7 @@ async def upload_fileobj(
     executor: Optional[ThreadPoolExecutor] = None,
     retries: int = 3,
     backoff_factor: float = 1.0,
-    callback_func: Optional[Callable[[int, int], Awaitable[None]]] = None
+    callback_func: Optional[Callable[[int, int], Awaitable[None]]] = None,
 ) -> int:
     """Upload a file object to a url.
 
@@ -403,7 +426,8 @@ async def upload_fileobj(
         chunk_size: The size of the chunk in the streaming upload. Note that this
             function does not do multipart upload. Default: 16MB.
         timeout: The total timeout of the upload in seconds.
-            By default, there is no total timeout, but only socket timeouts of 5s.
+            By default, there is no total timeout, but only socket connect timeout
+            of 5 seconds and a socket read timeout of 10 minutes.
         connector: An optional aiohttp connector to support connection pooling.
         md5: The MD5 digest (binary) of the file. Supply the MD5 if you already
             have access to it. Otherwise this function will compute it for you.
@@ -463,7 +487,7 @@ async def upload_fileobj(
             "PUT",
             url,
             headers=headers,
-            timeout=DEFAULT_TIMEOUT if timeout is None else timeout,
+            timeout=DEFAULT_UPLOAD_TIMEOUT if timeout is None else timeout,
         )
         response = await _request_with_retry(request, retries, backoff_factor)
         if response.status != 200: